Move alerts modules to src/alerts
