+++ src/alerts/Alerts.test.js
@@ -1,13 +1,14 @@
 import { fireEvent, getByText } from '@testing-library/react';
 import _ from 'lodash/fp';
 import { h } from 'react-hyperscript-helpers';
-import Alerts from 'src/components/Alerts';
-import { useServiceAlerts } from 'src/libs/service-alerts';
 import * as Utils from 'src/libs/utils';
 import { renderWithAppContexts as render } from 'src/testing/test-utils';
 
-jest.mock('src/libs/service-alerts', () => {
-  const originalModule = jest.requireActual('src/libs/service-alerts');
+import Alerts from './Alerts';
+import { useServiceAlerts } from './service-alerts';
+
+jest.mock('./service-alerts', () => {
+  const originalModule = jest.requireActual('./service-alerts');
   return {
     ...originalModule,
     useServiceAlerts: jest.fn(),
