+++ src/alerts/service-alerts-polling.js
@@ -1,5 +1,6 @@
 import _ from 'lodash/fp';
-import { getServiceAlerts, serviceAlertsStore } from 'src/libs/service-alerts';
+
+import { getServiceAlerts, serviceAlertsStore } from './service-alerts';
 
 export const startPollingServiceAlerts = () => {
   const loadServiceAlerts = () => getServiceAlerts().then((serviceAlerts) => serviceAlertsStore.set(serviceAlerts), _.noop);
