+++ src/appLoader.js
@@ -5,10 +5,10 @@ import _ from 'lodash/fp';
 import { createRoot } from 'react-dom/client';
 import { h } from 'react-hyperscript-helpers';
 import RModal from 'react-modal';
+import { startPollingServiceAlerts } from 'src/alerts/service-alerts-polling';
 import { initializeClientId } from 'src/auth/oidc-broker';
 import { initializeAuth } from 'src/libs/auth';
 import { isAxeEnabled } from 'src/libs/config';
-import { startPollingServiceAlerts } from 'src/libs/service-alerts-polling';
 import { initializeTCell } from 'src/libs/tcell';
 import Main from 'src/pages/Main';
